backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
bleach==6.2.0
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.2
django-chartjs==2.3.0
django-cors-headers==4.5.0
django-elasticsearch-dsl==8.0
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-model-utils==5.0.0
django-mptt==0.16.0
django-nested-admin==4.1.1
django-redis==5.4.0
django-simple-history==3.7.0
django-suit==0.2.28
django-summernote==0.8.20.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
et_xmlfile==2.0.0
filelock==3.16.1
gdown==5.2.0
greenlet==3.1.1
idna==3.10
isort==5.13.2
kombu==5.5.4
Markdown==3.5.2
markdownify==1.1.0
martor==1.6.45
mypy-extensions==1.0.0
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
playwright==1.48.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
pyee==12.0.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-monkey-business==1.1.0
razorpay==1.4.2
redis==5.2.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
schedule==1.2.2
setuptools==80.3.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stripe==11.5.0
tablib==3.7.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
